Move `_raise_wrapfail` to callers.py
It's only used there.
